+++ packages/redux-devtools-cli/src/bin/redux-devtools.ts
@@ -88,10 +88,9 @@ if (argv.injectserver) {
 }
 
 // eslint-disable-next-line @typescript-eslint/no-floating-promises
-server(argv).then(function (r) {
+server(argv).then(async function (r) {
   if (argv.open && argv.open !== 'false') {
-    r.on('ready', async function () {
-      await openApp(argv.open as string, options);
-    });
+    await r.listener('ready').once();
+    await openApp(argv.open as string, options);
   }
 });