+++ packages/redux-devtools-inspector-monitor-test-tab/demo/src/getOptions.ts
@@ -9,12 +9,12 @@ export default function getOptions(location: { search: string }) {
   return {
     useExtension: location.search.indexOf('ext') !== -1,
     supportImmutable: location.search.indexOf('immutable') !== -1,
-    theme: getTheme(),
+    theme: getTheme(location),
     dark: location.search.indexOf('dark') !== -1,
   };
 }
 
-function getTheme() {
+function getTheme(location: { search: string }) {
   const match = /theme=([^&]+)/.exec(location.search);
   return match ? match[1] : 'inspector';
 }