+++ packages/redux-devtools-rtk-query-monitor/demo/src/features/posts/PostsManager.tsx
@@ -17,7 +17,7 @@ import {
   StatNumber,
   useToast,
 } from '@chakra-ui/react';
-import { Route, Switch, useHistory } from 'react-router-dom';
+import { Route, Routes, useNavigate } from 'react-router-dom';
 import { MdBook } from 'react-icons/md';
 import React, { useState } from 'react';
 import {
@@ -88,7 +88,7 @@ const AddPost = () => {
 
 const PostList = () => {
   const { data: posts, isLoading } = useGetPostsQuery();
-  const history = useHistory();
+  const navigate = useNavigate();
 
   if (isLoading) {
     return <div>Loading</div>;
@@ -101,7 +101,7 @@ const PostList = () => {
   return (
     <List spacing={3}>
       {posts.map(({ id, name }) => (
-        <ListItem key={id} onClick={() => history.push(`/posts/${id}`)}>
+        <ListItem key={id} onClick={() => navigate(`/posts/${id}`)}>
           <ListIcon as={MdBook} color="green.500" /> {name}
         </ListItem>
       ))}
@@ -147,14 +147,17 @@ export const PostsManager = () => {
           </Box>
         </Box>
         <Box flex={2}>
-          <Switch>
-            <Route path="/posts/:id" component={PostDetail} />
-            <Route>
-              <Center h="200px">
-                <Heading size="md">Select a post to edit!</Heading>
-              </Center>
-            </Route>
-          </Switch>
+          <Routes>
+            <Route path="posts/:id" element={<PostDetail />} />
+            <Route
+              index
+              element={
+                <Center h="200px">
+                  <Heading size="md">Select a post to edit!</Heading>
+                </Center>
+              }
+            />
+          </Routes>
         </Box>
       </Flex>
     </Box>