+++ packages/redux-devtools-rtk-query-monitor/demo/src/features/posts/PostsView.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Switch, Route } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import { PostsManager } from '../../features/posts/PostsManager';
 import { Box, Heading } from '@chakra-ui/react';
 
@@ -7,9 +7,9 @@ function PostsView() {
   return (
     <Box as="section" p="2">
       <Heading as="h2">Posts Demo</Heading>
-      <Switch>
-        <Route path="/" component={PostsManager} />
-      </Switch>
+      <Routes>
+        <Route path="/*" element={<PostsManager />} />
+      </Routes>
     </Box>
   );
 }