+++ packages/redux-devtools-slider-monitor/examples/todomvc/src/index.tsx
@@ -1,10 +1,11 @@
 import 'todomvc-app-css/index.css';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import configureStore from './store/configureStore';
 import Root from './containers/Root';
 
 const store = configureStore();
 
 const rootEl = document.getElementById('root');
-ReactDOM.render(<Root store={store} />, rootEl);
+const root = ReactDOM.createRoot(rootEl!);
+root.render(<Root store={store} />);