+++ packages/redux-devtools-inspector-monitor-test-tab/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 1.0.0
+
+### Minor Changes
+
+- 8a7eae4: Add React 18 to peerDependencies range
+
+### Patch Changes
+
+- Updated dependencies [8a7eae4]
+  - @redux-devtools/inspector-monitor@3.0.0
+  - @redux-devtools/ui@1.3.0
+
+## 0.8.6
+
+### Patch Changes
+
+- @redux-devtools/inspector-monitor@2.1.2
+
 ## 0.8.0
 
 - Adds ESM build (https://github.com/reduxjs/redux-devtools/pull/997) and switches the default export to a named export in order to ensure that the CommonJS output and the ESM output are [interchangeable](https://rollupjs.org/guide/en/#outputexports):