+++ packages/redux-devtools-inspector-monitor/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Change Log
 
+## 3.0.1
+
+### Patch Changes
+
+- a55ba302: Fix peer dependencies on @redux-devtools/core
+- Updated dependencies [a55ba302]
+  - @redux-devtools/core@3.13.1
+
+## 3.0.0
+
+### Minor Changes
+
+- 8a7eae4: Add React 18 to peerDependencies range
+
+### Patch Changes
+
+- Updated dependencies [8a7eae4]
+  - react-json-tree@0.17.0
+  - @redux-devtools/core@3.13.0
+
+## 2.1.2
+
+### Patch Changes
+
+- Updated dependencies [4891bf6]
+  - @redux-devtools/core@3.12.0
+
 ## 2.1.1
 
 ### Patch Changes