+++ packages/redux-devtools-utils/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 2.0.1
+
+### Patch Changes
+
+- a55ba302: Fix peer dependencies on @redux-devtools/core
+- Updated dependencies [a55ba302]
+  - @redux-devtools/core@3.13.1
+
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [8a7eae4]
+  - @redux-devtools/core@3.13.0
+
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [4891bf6]
+  - @redux-devtools/core@3.12.0
+
 ## [1.0.0-6](https://github.com/reduxjs/redux-devtools/compare/@redux-devtools/utils@1.0.0-5...@redux-devtools/utils@1.0.0-6) (2021-03-06)
 
 **Note:** Version bump only for package @redux-devtools/utils